[jdbc->ebean] 重构原生 JDBC 为 Ebean 框架

+++ src/main/java/client/MapleQuestStatus.java
@@ -72,7 +72,7 @@ public class MapleQuestStatus implements Serializable {
         return quest;
     }
 
-    public final byte getStatus() {
+    public final int getStatus() {
         return status;
     }
