+++ src/main/java/server/StructRewardItem.java
@@ -4,6 +4,6 @@ public class StructRewardItem {
 
     public int itemid;
     public long period;
-    public short prob, quantity;
+    public int prob, quantity;
     public String effect, worldmsg;
 }
